Added object tile, working on player facing direction

+++ network/GateTile.h
@@ -11,12 +11,12 @@ public:
 	
 	// Default constructor
 	GateTile(std::vector<Key> aKeys = {}, int num = 0, int aWallLayout = 0, int aHeight = 0) :
-		Tile(TileType::GATE, aWallLayout, aHeight), gateNum(num), keyProgress(0), currentConstructTime(0.0f), validKeys(aKeys) {}
+		Tile(TileType::GATE, aWallLayout, aHeight), gateNum(num), keyProgress(0), currentConstructTime(0.0), validKeys(aKeys) {}
 
 	// Getters
 	int			getGateNum() { return gateNum; }
 	int			getKeyProgress() { return keyProgress; }
-	float		getCurrentConstructTime() { return currentConstructTime; }
+	double		getCurrentConstructTime() { return currentConstructTime; }
 	bool		hasAllKeys() { return keyProgress == 3; }
 	bool		isOpen() { return currentConstructTime >= TOTAL_CONSTRUCT_TIME; }
